Refactor the api, make straightforward
By recommendation of some other cryptographic engineers, I
refactored the api such that there is for the user of the crate
no specialization possible anymore.  All SigningKey, VerifyingKey,
etc. types and functions will be split.

We will still implement the RustCrypto trait collection on each
of them.  However, we will not expose any crate-internal traits
or generics structure to the user.

This is done because it makes it much easier to keep the public
api both straightforward, and this in turn makes it easier to
keep the public api stable.